requirements.txt
altgraph==0.16.1
certifi==2019.3.9
chardet==3.0.4
comtypes==1.1.7
future==0.17.1
idna==2.8
macholib==1.11
pefile==2019.4.18
PyInstaller==3.4
pywin32==224
pywin32-ctypes==0.2.0
requests==2.21.0
Unipath==1.1
urllib3==1.24.1
xlwings==0.15.5
